refactor: clarify features exposed by ironrdp-tls
